pyproject.toml
[tool.poetry]
name = "retina-therm"
version = "0.1.0"
description = ""
authors = ["CD Clark III <clifton.clark@gmail.com>"]
readme = "README.md"
packages = [{include = "retina_therm"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
mpmath = "^1.3.0"
rich = "^13.6.0"
typer = "^0.9.0"
pyyaml = "^6.0.1"
numpy = "^1.26.1"
scipy = "^1.11.3"
pint = "^0.22"
fspathtree = "^0.8"
p-tqdm = "^1.4.0"
pydantic = "^2.5.3"
tissue-properties = "^0.1.0"
tinydb = "^4.8.0"
wasmer = "^1.1.0"
wasmer-compiler-llvm = "^1.1.0"
wasmer-compiler-cranelift = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pudb = "^2023.1"
pytest-benchmark = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
retina-therm = "retina_therm.cli:app"
